+++ src/common/components/Logo.tsx
@@ -1,5 +1,5 @@
 import { Image, Box, ImageProps } from '@chakra-ui/react';
-import logo from '../static/images/logo.png';
+import logo from '../../static/images/logo.png';
 
 const Logo: React.FC<ImageProps> = (imageProps) => (
   <Box>
