+++ src/common/components/index.ts
@@ -1,2 +1,3 @@
 export * from './TextControl';
 export * from './Button';
+export * from './Logo';
