+++ src/common/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './capitalize';
+export * from './random-string';
