+++ src/common/utils/random-string.ts
@@ -1,0 +1,1 @@
+export const randomString = () => Math.random().toString(36).substr(3, 5);
