--- src/components/index.ts
@@ -1,4 +1,0 @@
-export * from './Register';
-export * from './Login';
-export * from './Navbar';
-export * from './Logo';
