+++ src/index.tsx
@@ -1,11 +1,14 @@
 import ReactDOM from 'react-dom';
-import App from './App';
-import { store } from './store';
 import { Provider } from 'react-redux';
 import { ChakraProvider } from '@chakra-ui/provider';
 import { BrowserRouter } from 'react-router-dom';
-import theme from './theme';
-import './theme/styles.css';
+
+import App from './app/App';
+import { store } from './app/store';
+import theme from './app/theme';
+
+import './app/theme/styles.css';
+
 const rootElement = document.getElementById('root');
 
 ReactDOM.render(
