--- src/store/index.ts
@@ -1,2 +1,0 @@
-export { store } from './store';
-export * from './hooks';
